+++ IISU/ClientPSCertStoreManager.cs
@@ -15,9 +15,11 @@ using Keyfactor.Logging;
 using Keyfactor.Orchestrators.Common.Enums;
 using Keyfactor.Orchestrators.Extensions;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.FlowAnalysis;
 using Microsoft.Extensions.Logging;
 using System;
 using System.IO;
+using System.Linq.Expressions;
 using System.Management.Automation;
 using System.Management.Automation.Runspaces;
 using System.Security.Cryptography.X509Certificates;
@@ -48,6 +50,10 @@ namespace Keyfactor.Extensions.Orchestrator.WindowsCertStore
 
         public string CreatePFXFile(string certificateContents, string privateKeyPassword)
         {
+            _logger.LogTrace("Entering CreatePFXFile");
+            if (!string.IsNullOrEmpty(privateKeyPassword)) { _logger.LogTrace("privateKeyPassword was present"); } 
+            else _logger.LogTrace("No privateKeyPassword Presented");
+
             try
             {
                 // Create the x509 certificate
@@ -79,11 +85,13 @@ namespace Keyfactor.Extensions.Orchestrator.WindowsCertStore
                     var results = ps.Invoke();
 
                     // Get the result (temporary file path) returned by the script
+                    _logger.LogTrace($"Results after creating PFX File: {results[0].ToString()}");
                     return results[0].ToString();
                 }
             }
-            catch (Exception)
+            catch (Exception ex)
             {
+                _logger.LogError(ex.ToString());
                 throw new Exception("An error occurred while attempting to create and write the X509 contents.");
             }
         }
@@ -108,55 +116,137 @@ namespace Keyfactor.Extensions.Orchestrator.WindowsCertStore
             }
         }
 
-        public JobResult ImportPFXFile(string filePath, string privateKeyPassword, string cryptoProviderName)
+        public JobResult ImportPFXFile(string filePath, string privateKeyPassword, string cryptoProviderName, string storePath)
         {
             try
             {
+                _logger.LogTrace("Entering ImportPFX");
+
                 using (PowerShell ps = PowerShell.Create())
                 {
                     ps.Runspace = _runspace;
 
                     if (cryptoProviderName == null)
                     {
-                        string script = @"
-                        param($pfxFilePath, $privateKeyPassword, $cspName)
-                        $output = certutil -importpfx -p $privateKeyPassword $pfxFilePath 2>&1
-                        $c = $LASTEXITCODE
-                        $output
-                        ";
+                        if (privateKeyPassword == null)
+                        {
+                            // If no private key password is provided, import the pfx file directory to the store using addstore argument
+                            string script = @"
+                            param($pfxFilePath, $storePath)
+                            $output = certutil -addstore $storePath $pfxFilePath 2>&1
+                            $exit_message = ""LASTEXITCODE:$($LASTEXITCODE)""
 
-                        ps.AddScript(script);
-                        ps.AddParameter("pfxFilePath", filePath);
-                        ps.AddParameter("privateKeyPassword", privateKeyPassword);
+                            if ($output.GetType().Name -eq ""String"")
+                            {
+                                $output = @($output, $exit_message)
+                            }
+                            else
+                            {
+                                $output += $exit_message
+                            }
+                            $output
+                            ";
+
+                            ps.AddScript(script);
+                            ps.AddParameter("pfxFilePath", filePath);
+                            ps.AddParameter("storePath", storePath);
+                        }
+                        else
+                        {
+                            // Use ImportPFX to import the pfx file with private key password to the appropriate cert store
+
+                            string script = @"
+                            param($pfxFilePath, $privateKeyPassword)
+                            $output = certutil -importpfx -p $privateKeyPassword $storePath $pfxFilePath 2>&1
+                            $exit_message = ""LASTEXITCODE:$($LASTEXITCODE)""
+                            $stuff = certutil -dump
+                            if ($stuff.GetType().Name -eq ""String"")
+                            {
+                                $stuff = @($stuff, $exit_message)
+                            }
+                            else
+                            {
+                                $stuff += $exit_message
+                            }
+                            $output
+                            $stuff
+                            ";
+
+                            ps.AddScript(script);
+                            ps.AddParameter("pfxFilePath", filePath);
+                            ps.AddParameter("privateKeyPassword", privateKeyPassword);
+                            ps.AddParameter("storePath", storePath);
+                        }
                     }
                     else
                     {
-                        string script = @"
-                        param($pfxFilePath, $privateKeyPassword, $cspName)
-                        $output = certutil -importpfx -csp $cspName -p $privateKeyPassword $pfxFilePath 2>&1
-                        $c = $LASTEXITCODE
-                        $output
-                        ";
+                        if (privateKeyPassword == null)
+                        {
+                            string script = @"
+                            param($pfxFilePath, $cspName, $storePath)
+                            $output = certutil -csp $cspName -addstore $storePath $pfxFilePath 2>&1
+                            $exit_message = ""LASTEXITCODE:$($LASTEXITCODE)""
 
-                        ps.AddScript(script);
-                        ps.AddParameter("pfxFilePath", filePath);
-                        ps.AddParameter("privateKeyPassword", privateKeyPassword);
-                        ps.AddParameter("cspName", cryptoProviderName);
+                            $stuff = certutil -dump
+                            if ($stuff.GetType().Name -eq ""String"")
+                            {
+                                $stuff = @($stuff, $exit_message)
+                            }
+                            else
+                            {
+                                $stuff += $exit_message
+                            }
+                            $output
+                            $stuff
+                            ";
+
+                            ps.AddScript(script);
+                            ps.AddParameter("pfxFilePath", filePath);
+                            ps.AddParameter("cspName", cryptoProviderName);
+                            ps.AddParameter("storePath", storePath);
+                        }
+                        else
+                        {
+                            string script = @"
+                            param($pfxFilePath, $privateKeyPassword, $cspName)
+                            $output = certutil -importpfx -csp $cspName -p $privateKeyPassword $storePath $pfxFilePath 2>&1
+                            $exit_message = ""LASTEXITCODE:$($LASTEXITCODE)""
+
+                            $stuff = certutil -dump
+                            if ($stuff.GetType().Name -eq ""String"")
+                            {
+                                $stuff = @($stuff, $exit_message)
+                            }
+                            else
+                            {
+                                $stuff += $exit_message
+                            }
+                            $output
+                            $stuff
+                            ";
+
+                            ps.AddScript(script);
+                            ps.AddParameter("pfxFilePath", filePath);
+                            ps.AddParameter("privateKeyPassword", privateKeyPassword);
+                            ps.AddParameter("cspName", cryptoProviderName);
+                            ps.AddParameter("storePath", storePath);
+                        }
                     }
 
                     // Invoke the script
+                    _logger.LogTrace("Attempting to import the PFX");
                     var results = ps.Invoke();
 
                     // Get the last exist code returned from the script
-                    // This statement is in a try/catch block because PSVariable.GetValue() is not a valid method on a remote PS Session and throws an exception.
-                    // Due to security reasons and Windows architecture, retreiving values from a remote system is not supported.
                     int lastExitCode = 0;
                     try
                     {
-                        lastExitCode = (int)ps.Runspace.SessionStateProxy.PSVariable.GetValue("c");
+                        lastExitCode = GetLastExitCode(results[^1].ToString());
+                        _logger.LogTrace($"Last exit code: {lastExitCode}");
                     }
                     catch (Exception)
                     {
+                        _logger.LogTrace("Unable to get the last exit code.");
                     }
                     
 
@@ -182,7 +272,10 @@ namespace Keyfactor.Extensions.Orchestrator.WindowsCertStore
                         foreach (var result in results)
                         {
                             string outputLine = result.ToString();
-                            if (!string.IsNullOrEmpty(outputLine) && outputLine.Contains("Error"))
+
+                            _logger.LogTrace(outputLine);
+
+                            if (!string.IsNullOrEmpty(outputLine) && outputLine.Contains("Error") || outputLine.Contains("permissions are needed"))
                             {
                                 isError = true;
                                 _logger.LogError(outputLine);
@@ -215,6 +308,30 @@ namespace Keyfactor.Extensions.Orchestrator.WindowsCertStore
                     JobHistoryId = _jobNumber,
                     FailureMessage = $"Error Occurred in ImportPFXFile {LogHandler.FlattenException(e)}"
                 };
+            }
+        }
+
+        private int GetLastExitCode(string result)
+        {
+            // Split the string by colon
+            string[] parts = result.Split(':');
+
+            // Ensure the split result has the expected parts
+            if (parts.Length == 2 && parts[0] == "LASTEXITCODE")
+            {
+                // Parse the second part into an integer
+                if (int.TryParse(parts[1], out int lastExitCode))
+                {
+                    return lastExitCode;
+                }
+                else
+                {
+                    throw new Exception("Failed to parse the LASTEXITCODE value.");
+                }
+            }
+            else
+            {
+                throw new Exception("The last element does not contain the expected format.");
             }
         }